Move internal.lock_machines to its own module
